issue-296 Create basic new:role cli command

+++ packages/cli/src/services/generator.ts
@@ -6,5 +6,10 @@ import { Target } from './generator/target'
 export async function generate<TInfo>(target: Target<TInfo>): Promise<void> {
   const rendered = Mustache.render(target.template, target.info)
   const renderPath = path.join(process.cwd(), target.placementDir, `${target.name}${target.extension}`)
-  await fs.outputFile(renderPath, rendered)
+
+  if (target.append && fs.existsSync(renderPath)) {
+    await fs.appendFile(renderPath, rendered)
+  } else {
+    await fs.outputFile(renderPath, rendered)
+  }
 }
